requirements.txt
# Core dependencies
firecrawl-py>=0.0.16
anthropic>=0.18.0
pyyaml>=6.0
requests>=2.31.0

# Database
sqlite3  # Built-in with Python

# CLI
click>=8.1.0
rich>=13.0.0
tabulate>=0.9.0

# Utilities
python-dotenv>=1.0.0
schedule>=1.2.0
